+++ packages/student-witness/src/index.tsx
@@ -1,8 +1,3 @@
-// import get from 'lodash/get';
-// import { lib } from './utils/lib';
-// import '#/components/global';
-// import './assets/style.less';
-// import { c } from '#/components/utils';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import Icon from '@/assets/more.svg';
@@ -24,7 +19,7 @@ class App extends React.Component {
     return (
       <Page>
         <h1>优秀学员故事</h1>
-        <Icon width="37" height="35" />
+        <Icon width="19" height="18" />
       </Page>
     )
   }