+++ packages/student-witness/typings/index.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="../../../typings/styled-components.d.ts" />
+/// <reference path="../../../typings/svgr-webpack.d.ts" />